676 - Don’t allow deleting project instance if there are executions

+++ server/libs/automation/automation-configuration/automation-configuration-service/src/main/java/com/bytechef/automation/configuration/facade/ProjectInstanceFacadeImpl.java
@@ -156,6 +156,11 @@ public class ProjectInstanceFacadeImpl implements ProjectInstanceFacade {
                         projectInstanceWorkflow.getId()));
             }
 
+            if (OptionalUtils.isPresent(instanceJobService.fetchLastJobId(id, Type.AUTOMATION))) {
+                throw new IllegalStateException(
+                    "The ProjectInstance id=%s has executed workflows.".formatted(id));
+            }
+
             projectInstanceWorkflowService.delete(projectInstanceWorkflow.getId());
         }
